% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{geocode}
\alias{geocode}
\title{Geocode gets the longitude and latitude of a given location.}
\usage{
geocode(location)
}
\arguments{
\item{location}{String that describes a geographic location.}
}
\value{
A data frame with the input string (location), the longitude and the latitude.
}
\description{
The function \emph{geocode} retrieves the longitude and the latitude of a location through www.openstreetmap.org.
}
\examples{
long_lat <- geocode("Amsterdam, Netherlands")

}
