% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_icon}
\alias{get_icon}
\title{get_icon gets weather icons.}
\usage{
get_icon(location, api_key)
}
\arguments{
\item{location}{String that describes a geographic location.}

\item{api_key}{String that represents your personal api Key from www.openweathermap.org}
}
\value{
An image of the weather icons.
}
\description{
The function \emph{get_icon} saves the icon that reflects the current weather.
}
