% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_icon_map}
\alias{get_icon_map}
\title{get_icon_map creates a map with weather icons.}
\usage{
get_icon_map(location, api_key)
}
\arguments{
\item{location}{String that describes a geographic location.}

\item{api_key}{String that represents your personal api Key from www.openweathermap.org}
}
\value{
An image that consists of a map with weather icons.
}
\description{
The function \emph{get_icon_map} creates a map of a location with weather icons incorporated in it. The icons reflect the current weather.
}
