% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_map}
\alias{get_map}
\title{get_map gets and saves a map using leaflet()}
\usage{
get_map(location)
}
\arguments{
\item{location}{String that describes a geographic location.}
}
\value{
A png of a map.
}
\description{
get_map gets and saves a map using leaflet()
}
\examples{
my_map <- get_map("Amsterdam, Niederlande")

}
