% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_weather_gif}
\alias{get_weather_gif}
\title{Creates a weather GIF for the current weather at a specific location.}
\usage{
get_weather_gif(location, apiKey)
}
\arguments{
\item{location}{String that describes a geographic location.}

\item{apiKey}{String that represents your personal api Key from www.openweathermap.org}
}
\value{
A gif that reflects the weather at a chosen location.
}
\description{
Creates a weather GIF for the current weather at a specific location.
}
