% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_weather}
\alias{get_weather}
\title{Retrieves weather data for an input location}
\usage{
get_weather(location, apiKey)
}
\arguments{
\item{location}{String that describes a geographical location.}

\item{apiKey}{String that represents your personal apiKey. To receive an apiKey, please visit www.openweathermap.org and sign up for free.}
}
\value{
A complex list that includes a variety of weather data, such as current weather, daily weather, ...
}
\description{
The function \emph{get_weather} retrieves weatherdata for the input location via www.openweathermap.org
}
