% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_app_functions.R
\name{get_weather_image}
\alias{get_weather_image}
\title{get_weather_image returns an image that reflects the current weather}
\usage{
get_weather_image(location, api_key)
}
\arguments{
\item{location}{String that describes a geographic location.}

\item{api_key}{String that represents your personal api Key from www.openweathermap.org}
}
\value{
An image that reflects the current weahter forecast of a location (Either clear sky, rainy, cloudy, storm, misty, snowy)
}
\description{
get_weather_image returns an image that reflects the current weather
}
